refactor(#116): change name
