+++ src/chat/dto/response-find-chat-rooms-pagination.dto.ts
@@ -1,21 +1,21 @@
 import { PaginationResponseDto } from 'src/common/dto/pagination-response.dto';
-import { ResponseGetChatRoomsDto } from './response-get-chat-rooms.dto';
+import { ResponseFindChatRoomsDto } from './response-find-chat-rooms.dto';
 import { ApiProperty } from '@nestjs/swagger';
 
-export class ResponseGetChatRoomsPaginationDto extends PaginationResponseDto {
+export class ResponseFindChatRoomsPaginationDto extends PaginationResponseDto {
   @ApiProperty({
     description: '최신 채팅 내역이 포함된 chatRoom 객체',
-    type: [ResponseGetChatRoomsDto],
+    type: [ResponseFindChatRoomsDto],
   })
-  chatRooms: ResponseGetChatRoomsDto[];
+  chatRooms: ResponseFindChatRoomsDto[];
 
   constructor(
-    responseGetChatRoomsDto: ResponseGetChatRoomsDto[],
+    responseFindChatRoomsDto: ResponseFindChatRoomsDto[],
     totalCount: number,
     page: number,
     pageSize: number,
   ) {
     super(totalCount, page, pageSize);
-    this.chatRooms = responseGetChatRoomsDto;
+    this.chatRooms = responseFindChatRoomsDto;
   }
 }
