+++ src/chat/dto/response-find-chat-rooms.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { ChatUserDto } from 'src/users/dtos/chat-user.dto';
 import { AggregateChatRoomsDto } from './aggregate-chat-rooms.dto';
 
-export class ResponseGetChatRoomsDto {
+export class ResponseFindChatRoomsDto {
   @ApiProperty({
     description: '최신 채팅 내역이 포함된 chatRoom 객체',
   })
