+++ src/chat/swagger-decorators/find-one-chat-room-by-user-id.decorator.ts
@@ -2,41 +2,26 @@ import { applyDecorators } from '@nestjs/common';
 import {
   ApiExtraModels,
   ApiHeaders,
-  ApiOkResponse,
   ApiOperation,
-  ApiParam,
   ApiResponse,
   getSchemaPath,
 } from '@nestjs/swagger';
-import { AggregateChatRoomForChatsDto } from '../dto/aggregate-chat-room-for-chats.dto';
+import { ChatRoomDto } from '../dto/chat-room.dto';
 
-export function ApiGetChats() {
+export function ApiFindOneChatRoomByUserId() {
   return applyDecorators(
     ApiOperation({
-      summary: '해당 채팅룸 채팅 전체 조회',
-      description: 'Param - room-id, Headers - access_token',
+      summary: '유저 id로 채팅룸 단일 조회',
+      description: 'Param - roomId',
     }),
-    ApiOkResponse({
-      description:
-        '성공적으로 채팅방 채팅 조회 및 읽지 않았던 채팅들 seenUsers에 myId 추가 및 pagination함. 20개씩 잘라서 가져옴.',
+    ApiResponse({
+      status: 200,
+      description: '성공적으로 채팅방 (단일)조회',
       schema: {
         properties: {
           content: {
             type: 'object',
-            $ref: getSchemaPath(AggregateChatRoomForChatsDto),
-          },
-        },
-      },
-    }),
-    ApiResponse({
-      status: 400,
-      description: '유효성 검사 실패',
-      content: {
-        JSON: {
-          example: {
-            message: '올바른 ObjectId 형식이 아닙니다.',
-            error: 'Bad Request',
-            statusCode: 400,
+            $ref: getSchemaPath(ChatRoomDto),
           },
         },
       },
@@ -52,17 +37,13 @@ export function ApiGetChats() {
     }),
     ApiResponse({
       status: 403,
-      description:
-        '만료된 액세스 토큰인 경우, 해당 채팅방에 접근 권한 없는 경우',
+      description: '만료된 액세스 토큰인 경우',
       content: {
         JSON: {
           example: {
             statusCode: 403,
             error: 'Forbidden',
-            message: [
-              '만료된 토큰입니다.',
-              '해당 채팅방에 접근 권한이 없습니다',
-            ],
+            message: '만료된 토큰입니다.',
           },
         },
       },
@@ -97,13 +78,6 @@ export function ApiGetChats() {
         example: '여기에 액세스 토큰',
       },
     ]),
-    ApiParam({
-      name: 'roomId',
-      description: '채팅방의 id',
-      required: true,
-      type: 'string',
-      format: 'objectId',
-    }),
-    ApiExtraModels(AggregateChatRoomForChatsDto),
+    ApiExtraModels(ChatRoomDto),
   );
 }
